fix: split Blog template into components
